allow keys like KeyGamepadLStickUp to overlap (#20)
When D-pad emulating stick enters a diagonal zone,
both directions will be implied as activated due to the
overlapping areas.

This makes it possible to fully emulate the D-pad controls
using the stick. For instance, the player can press both
left and up D-pad buttons and get the diagonal movement work.
It was not possible to achieve that with a stick before.