add a way to configure gamepad sticks deadzone
The `Handler.GamepadDeadzone` specifies the input magnitude threshold.
The default value is `0.055`, just as before, but not it's possible
to re-configure it for every handler.

For instance, one handler (1st player gamepad) could have a deadzone
set to 0.1 because it's flaky while the second gamepad could remain
on 0.055 setting.

It's up to the game developer to export the ability to configure the
deadzone thresholds for the players.