implement ActionIsJustReleased method
This is not the final implementation as it doesn't handle
some tricky combinations like joystick movements that emulate a D-pad.

It also ignores the taps as they're activated at the moment the gesture
is finished (i.e. the finger is detached from a screen).

Right now we can handle gamepad, keyboard and mouse key release events.

Key modifiers work too and it's not necessary to release the modifier
keys during the same frame as the main key.
So, a `ctrl+lmb` released event would trigger if `lmb` is released
while `ctrl` is still being pressed.

A new example demonstrates how to use this new feature.
It also demonstrates how to avoid some of the conflicts
described in #36 (this is not a novel issue and it's not related
to the release events directly).

Refs #25